@@ -4,7 +4,7 @@ return [
     'dependencies' => [
         'factories' => [
             // Logger using the default keys.
-            \Psr\Log\LoggerInterface::class => \WShafer\PSR11MonoLog\MonologFactory::class,
+            \Psr\Log\LoggerInterface::class => \Blazon\PSR11MonoLog\MonologFactory::class,
         ]
     ],
 
