@@ -13,6 +13,8 @@ $cacheConfig = [
 ];
 
 $aggregator = new ConfigAggregator([
+    \Mezzio\Flash\ConfigProvider::class,
+    \Laminas\Diactoros\ConfigProvider::class,
     \Laminas\Cache\ConfigProvider::class,
     \Laminas\Form\ConfigProvider::class,
     \Laminas\InputFilter\ConfigProvider::class,
@@ -23,8 +25,6 @@ $aggregator = new ConfigAggregator([
     \Mezzio\Authentication\ConfigProvider::class,
     \Mezzio\Session\ConfigProvider::class,
     \Mezzio\Csrf\ConfigProvider::class,
-    \WShafer\PSR11MonoLog\ConfigProvider::class,
-
     \Laminas\HttpHandlerRunner\ConfigProvider::class,
     \Mezzio\Router\FastRouteRouter\ConfigProvider::class,
 
