@@ -2,8 +2,8 @@
 
 declare(strict_types = 1);
 
-use Zend\DI\Config\Config;
-use Zend\DI\Config\ContainerFactory;
+use Elie\PHPDI\Config\Config;
+use Elie\PHPDI\Config\ContainerFactory;
 
 $config  = require __DIR__ . '/config.php';
 
