@@ -11,6 +11,7 @@ use DateTime;
 use Exception;
 use Fig\Http\Message\StatusCodeInterface;
 use GuzzleHttp\Psr7\Response;
+use JSHayes\FakeRequests\RequestHandler;
 use PHPUnit\Framework\AssertionFailedError;
 use Psr\Http\Message\RequestInterface;
 
@@ -31,6 +32,10 @@ class LpaContext implements Context
 {
     use ActorContext;
     use BaseUiContextTrait;
+
+    /** @var RequestHandler Allows the overriding of the dashboard LPA endpoints request (if registered) */
+    private RequestHandler $requestDashboardLPAs;
+    private RequestHandler $requestDashboardLPACodes;
 
     /**
      * @Then /^I am taken to a page explaining why instructions and preferences are not available$/
@@ -224,6 +229,15 @@ class LpaContext implements Context
     public function iAmOnTheAddAnLPATriagePage()
     {
         $this->ui->visit('/lpa/add');
+        $this->iAmTakenToTheAddAnLPATriagePage();
+    }
+
+    /**
+     * @Then /^I am taken to the add an LPA triage page$/
+     */
+    public function iAmTakenToTheAddAnLPATriagePage()
+    {
+        $this->ui->assertPageAddress('/lpa/add');
         $this->ui->assertPageContainsText('Do you have an activation key to add an LPA?');
     }
 
@@ -249,26 +263,6 @@ class LpaContext implements Context
      */
     public function iAmOnTheDashboardPage()
     {
-        //API call for getting all the users added LPAs
-        $this->apiFixtures->get('/v1/lpas')
-            ->respondWith(
-                new Response(
-                    StatusCodeInterface::STATUS_OK,
-                    [],
-                    json_encode([$this->userLpaActorToken => $this->lpaData])
-                )
-            );
-
-        //API call for getting each LPAs share codes
-        $this->apiFixtures->get('/v1/lpas/' . $this->userLpaActorToken . '/codes')
-            ->respondWith(
-                new Response(
-                    StatusCodeInterface::STATUS_OK,
-                    [],
-                    json_encode([])
-                )
-            );
-
         $this->ui->visit('/lpa/dashboard');
 
         $this->ui->assertResponseStatus(StatusCodeInterface::STATUS_OK);
@@ -356,7 +350,7 @@ class LpaContext implements Context
     }
 
     /**
-     * @Given /^I have added 2 LPAs to my account$/
+     * @Given /^I have added an additional LPA to my account$/
      */
     public function iHaveAdded2LPAsToMyAccount()
     {
@@ -380,8 +374,7 @@ class LpaContext implements Context
         }
 
         //API call for getting all the users added LPAs
-        $this->apiFixtures->get('/v1/lpas')
-            ->respondWith(
+        $this->requestDashboardLPAs->respondWith(
                 new Response(
                     StatusCodeInterface::STATUS_OK,
                     [],
@@ -1359,8 +1352,54 @@ class LpaContext implements Context
     }
 
     /**
-     * @Given I have been given access to use an LPA via credentials
      * @Given I have added an LPA to my account
+     */
+    public function iHaveAddedAnLpaToMyAccount()
+    {
+        $this->iHaveBeenGivenAccessToUseAnLPAViaCredentials();
+
+        //API call for getting all the users added LPAs
+        $this->requestDashboardLPAs = $this->apiFixtures->get('/v1/lpas')
+            ->respondWith(
+                new Response(
+                    StatusCodeInterface::STATUS_OK,
+                    [],
+                    json_encode([$this->userLpaActorToken => $this->lpaData])
+                )
+            );
+
+        //API call for getting each LPAs share codes
+        $this->requestDashboardLPACodes = $this->apiFixtures->get('/v1/lpas/' . $this->userLpaActorToken . '/codes')
+            ->respondWith(
+                new Response(
+                    StatusCodeInterface::STATUS_OK,
+                    [],
+                    json_encode([])
+                )
+            );
+    }
+
+    /**
+     * @Given I have no LPAs in my account
+     */
+    public function iHaveNoLpasInMyAccount()
+    {
+        $this->iHaveBeenGivenAccessToUseAnLPAViaCredentials();
+
+        //API call for getting all the users added LPAs
+        $this->apiFixtures->get('/v1/lpas')
+            ->respondWith(
+                new Response(
+                    StatusCodeInterface::STATUS_OK,
+                    [],
+                    json_encode([])
+                )
+            );
+    }
+
+    /**
+     * @Given I have been given access to use an LPA via credentials
+     *
      */
     public function iHaveBeenGivenAccessToUseAnLPAViaCredentials()
     {
@@ -1403,11 +1442,12 @@ class LpaContext implements Context
             'applicationHasRestrictions' => true,
             'applicationHasGuidance' => false,
             'lpa' => $this->lpa,
+            'added' => '2021-10-5 12:00:00'
         ];
     }
 
     /**
-     * @Given /^I have added an LPA to my account which has a donor signature before 2016$/
+     * @Given I have been given access to use an LPA via credentials which has a donor signature before 2016
      */
     public function iHaveBeenGivenAccessToUseAnLPAViaCredentialsSignedBefore2016()
     {
@@ -1454,16 +1494,63 @@ class LpaContext implements Context
             'lpaDonorSignatureDate' => '2015-06-30',
             'lpa' => $this->lpa,
         ];
+
+        //API call for getting all the users added LPAs
+        $this->requestDashboardLPAs = $this->apiFixtures->get('/v1/lpas')
+            ->respondWith(
+                new Response(
+                    StatusCodeInterface::STATUS_OK,
+                    [],
+                    json_encode([$this->userLpaActorToken => $this->lpaData])
+                )
+            );
+
+        //API call for getting each LPAs share codes
+        $this->requestDashboardLPACodes = $this->apiFixtures->get('/v1/lpas/' . $this->userLpaActorToken . '/codes')
+            ->respondWith(
+                new Response(
+                    StatusCodeInterface::STATUS_OK,
+                    [],
+                    json_encode([])
+                )
+            );
     }
 
+    /**
+     * @Given I have added an LPA to my account which has a donor signature before 2016
+     */
+    public function iHaveAddedAnLPAToMyAccountSignedBefore2016()
+    {
+        $this->iHaveBeenGivenAccessToUseAnLPAViaCredentialsSignedBefore2016();
+
+        //API call for getting all the users added LPAs
+        $this->requestDashboardLPAs = $this->apiFixtures->get('/v1/lpas')
+            ->respondWith(
+                new Response(
+                    StatusCodeInterface::STATUS_OK,
+                    [],
+                    json_encode([$this->userLpaActorToken => $this->lpaData])
+                )
+            );
+
+        //API call for getting each LPAs share codes
+        $this->requestDashboardLPACodes = $this->apiFixtures->get('/v1/lpas/' . $this->userLpaActorToken . '/codes')
+            ->respondWith(
+                new Response(
+                    StatusCodeInterface::STATUS_OK,
+                    [],
+                    json_encode([])
+                )
+            );
+    }
 
     /**
      * @Given /^I have created an access code$/
      */
     public function iHaveCreatedAnAccessCode()
     {
-        $this->iRequestToGiveAnOrganisationAccessToOneOfMyLPAs();
-        $this->iAmGivenAUniqueAccessCode();
+        $this->organisation = "TestOrg";
+        $this->accessCode = "XYZ321ABC987";
     }
 
     /**
@@ -1472,8 +1559,60 @@ class LpaContext implements Context
     public function iHaveGeneratedAnAccessCodeForAnOrganisationAndCanSeeTheDetails()
     {
         $this->iHaveCreatedAnAccessCode();
+        $this->iAmOnTheDashboardPage();
         $this->iClickToCheckMyAccessCodes();
         $this->iCanSeeAllOfMyAccessCodesAndTheirDetails();
+    }
+
+    /**
+     * @Given I have generated an access code for an LPA on my account
+     */
+    public function iHaveGeneratedAnAccessCodeForAnLPAOnMyAccount()
+    {
+        $this->iHaveBeenGivenAccessToUseAnLPAViaCredentials();
+
+        $this->organisation = "TestOrg";
+        $this->accessCode = "XYZ321ABC987";
+
+        // API call for get LpaById
+        $this->apiFixtures->get('/v1/lpas/' . $this->userLpaActorToken)
+            ->respondWith(
+                new Response(
+                    StatusCodeInterface::STATUS_OK,
+                    [],
+                    json_encode(
+                        [
+                            'user-lpa-actor-token' => $this->userLpaActorToken,
+                            'date' => 'date',
+                            'lpa' => $this->lpa,
+                            'actor' => $this->lpaData['actor'],
+                        ]
+                    )
+                )
+            );
+
+        // API call to get access codes
+        $this->apiFixtures->get('/v1/lpas/' . $this->userLpaActorToken . '/codes')
+            ->respondWith(
+                new Response(
+                    StatusCodeInterface::STATUS_OK,
+                    [],
+                    json_encode(
+                        [
+                            0 => [
+                                'SiriusUid' => $this->lpa->uId,
+                                'Added' => (new DateTime('yesterday'))->format('c'),
+                                'Expires' => (new DateTime('+1 month'))->setTime(23, 59, 59)->format('c'),
+                                'UserLpaActor' => $this->userLpaActorToken,
+                                'Organisation' => $this->organisation,
+                                'ViewerCode' => $this->accessCode,
+                                'Viewed' => false,
+                                'ActorId' => $this->actorId,
+                            ],
+                        ]
+                    )
+                )
+            );
     }
 
     /**
@@ -1800,8 +1939,6 @@ class LpaContext implements Context
      */
     public function iRequestToGiveAnOrganisationAccess()
     {
-        $this->iAmOnTheDashboardPage();
-
         // API call for get LpaById (when give organisation access is clicked)
         $this->apiFixtures->get('/v1/lpas/' . $this->userLpaActorToken)
             ->respondWith(
@@ -1819,9 +1956,8 @@ class LpaContext implements Context
                 )
             );
 
-        $this->ui->assertPageAddress('lpa/dashboard');
         $this->ui->clickLink('Give an organisation access');
-        $this->ui->assertPageAddress('lpa/code-make?lpa=' . $this->userLpaActorToken);
+        $this->ui->assertPageAddress('/lpa/code-make?lpa=' . $this->userLpaActorToken);
     }
 
     /**
@@ -2057,6 +2193,14 @@ class LpaContext implements Context
     public function iSelectToAddAnLPA()
     {
         $this->ui->clickLink('Add another LPA');
+    }
+
+    /**
+     * @When /^I choose to add my first LPA$/
+     */
+    public function iChooseToAddMyFirstLPA()
+    {
+        $this->ui->clickLink('Add your first LPA');
     }
 
     /**
@@ -2432,7 +2576,7 @@ class LpaContext implements Context
     }
 
     /**
-     * @When /^The status of the LPA got Revoked$/
+     * @When /^The LPA has been revoked$/
      * @Then /^I cannot see my access codes and their details$/
      */
     public function theStatusOfTheLpaGotRevoked()
