@@ -51,16 +51,17 @@ class SessionCsrfGuard implements CsrfGuardInterface
         $storedTokens = $this->session->get($csrfKey, []);
 
         $tokenParts = $this->splitHash($token);
-        $tokenId= $tokenParts['token'];
+        $tokenId = $tokenParts['token'];
 
         return array_key_exists($tokenId, $storedTokens)
             && $token === $this->formatHash($tokenId, $storedTokens[$tokenId]);
     }
 
-    protected function cleanupTokens(array $tokens): array {
+    protected function cleanupTokens(array $tokens): array
+    {
         $currentRequestId = $this->requestId;
 
-        return array_filter($tokens, function($requestId) use ($currentRequestId) {
+        return array_filter($tokens, function ($requestId) use ($currentRequestId) {
             return $requestId === $currentRequestId;
         });
     }
@@ -75,8 +76,8 @@ class SessionCsrfGuard implements CsrfGuardInterface
         $data = explode('-', $hash);
 
         return [
-            'token' => $data[0] ?: null,
-            'requestId' => $data[1] ?: null
+            'token' => $data[0] ?? null,
+            'requestId' => $data[1] ?? null
         ];
     }
 }
