@@ -8,10 +8,10 @@ use Actor\Form\LpaAdd;
 use Common\Form\AbstractForm;
 use Common\Form\Element\Csrf;
 use Common\Form\Fieldset\Date;
-use CommonTest\Form\{TestsLaminasForm, LaminasFormTests};
-use PHPUnit\Framework\TestCase;
-use Mezzio\Csrf\CsrfGuardInterface;
+use CommonTest\Form\{LaminasFormTests, TestsLaminasForm};
 use Laminas\Form\Element\Text;
+use Mezzio\Csrf\CsrfGuardInterface;
+use PHPUnit\Framework\TestCase;
 
 class LpaAddTest extends TestCase implements TestsLaminasForm
 {
@@ -40,7 +40,7 @@ class LpaAddTest extends TestCase implements TestsLaminasForm
         ];
     }
 
-    public function setUp()
+    public function setUp(): void
     {
         $guardProphecy = $this->prophesize(CsrfGuardInterface::class);
         $this->form = new LpaAdd($guardProphecy->reveal());
