@@ -35,7 +35,7 @@ class RemoveLpaTest extends TestCase implements TestsLaminasForm
         ];
     }
 
-    public function setUp()
+    public function setUp(): void
     {
         $guardProphecy = $this->prophesize(CsrfGuardInterface::class);
         $this->form = new RemoveLpa($guardProphecy->reveal());
