@@ -33,7 +33,7 @@ class CheckDetailsAndConsentTest extends TestCase implements TestsLaminasForm
         ];
     }
 
-    public function setUp()
+    public function setUp(): void
     {
         $guardProphecy = $this->prophesize(CsrfGuardInterface::class);
         $this->form = new CheckDetailsAndConsent($guardProphecy->reveal());
