@@ -34,7 +34,7 @@ class RequestPostcodeTest extends TestCase implements TestsLaminasForm
         ];
     }
 
-    public function setUp()
+    public function setUp(): void
     {
         $guardProphecy = $this->prophesize(CsrfGuardInterface::class);
         $this->form = new RequestPostcode($guardProphecy->reveal());
