@@ -20,7 +20,7 @@ class EmailClientTest extends TestCase
      */
     private $locale;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->notifyClientProphecy = $this->prophesize(NotifyClient::class);
         $this->locale = "en_GB";
