@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace CommonTest\Service\Lpa;
 
-use App\Exception\BadRequestException;
 use Common\Entity\CaseActor;
 use Common\Exception\ApiException;
 use Common\Service\ApiClient\Client as ApiClient;
@@ -589,7 +588,7 @@ class AddOlderLpaTest extends TestCase
 
     /**
      * @test
-     * @covers ::confir
+     * @covers ::confirm
      */
     public function it_will_fail_to_add_lpa_when_lpa_is_not_cleansed(): void
     {
