@@ -19,7 +19,7 @@ class SiriusLpaFactoryTest extends TestCase
     /** @var array */
     protected $simpleExampleFixtureData;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->fullExampleFixtureData = json_decode(file_get_contents(__DIR__ . '/../../../fixtures/full_example.json'), true);
         $this->simpleExampleFixtureData = json_decode(file_get_contents(__DIR__ . '/../../../fixtures/simple_example.json'), true);
