@@ -14,7 +14,7 @@ class DateValidatorTest extends TestCase
      */
     private $validator;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->validator = new DateValidator();
     }
